[#345] - Show charts with patterns
