+++ frontend/app/templates/components/journal-entry.hbs
@@ -8,4 +8,4 @@
   <div class="progressbar" style={{style}}></div>
 </div>
 
-{{link-to 'Edit' 'checkin' checkinId 'summary'}}
+{{link-to 'Edit' 'checkin.show' checkinId 'summary'}}
